+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The changelog is available [on GitHub][2].
 * [#41](https://github.com/kowainik/relude/issues/41):
   Support GHC-8.10.1.
   (by [@chshersh](https://github.com/chshersh))
+* [#24](https://github.com/kowainik/relude/issues/24):
+  Add `validationAll`, `when*` and `maybe*` combinators into
+  `Validation.Combinators`.
+  (by [@vrom911](https://github.com/vrom911))
 
 ## 0.0.0.0